+++ main.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"net/http"
 	_ "net/http/pprof"
 	"strings"
 	"time"
@@ -40,7 +38,7 @@ func main() {
 	utils.KeployVersion = version
 	fmt.Println(logo, " ")
 	fmt.Printf("version: %v\n\n", version)
-	latestVersion, err2 := getLatestGitHubRelease()
+	releaseInfo, err2 := utils.GetLatestGitHubRelease()
 	if err2 != nil {
 		log.Debug("Failed to fetch latest release version", err2)
 	}
@@ -53,8 +51,8 @@ func main() {
 	   │ Run %v to update         │
 	   ╰─────────────────────────────────────╯
 	`
-	versionmsg := fmt.Sprintf(msg, strings.TrimSpace(version), strings.TrimSpace(latestVersion), updatetext)
-	if latestVersion != version {
+	versionmsg := fmt.Sprintf(msg, strings.TrimSpace(version), strings.TrimSpace(releaseInfo.TagName), updatetext)
+	if releaseInfo.TagName != version {
 		fmt.Printf(versionmsg)
 	}
 
@@ -72,42 +70,4 @@ func main() {
 	defer utils.HandlePanic()
 	defer sentry.Flush(2 * time.Second)
 	cmd.Execute()
-}
-
-type GitHubRelease struct {
-	TagName string `json:"tag_name"`
-}
-
-// getLatestGitHubRelease fetches the latest version from GitHub releases.
-func getLatestGitHubRelease() (string, error) {
-	// GitHub repository details
-	repoOwner := "keploy"
-	repoName := "keploy"
-
-	// GitHub API URL for latest release
-	apiURL := fmt.Sprintf("https://api.github.com/repos/%s/%s/releases/latest", repoOwner, repoName)
-
-	// Create an HTTP client
-	client := http.Client{}
-
-	// Create a GET request to the GitHub API
-	req, err := http.NewRequest("GET", apiURL, nil)
-	if err != nil {
-		return "", err
-	}
-
-	// Send the request
-	resp, err := client.Do(req)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	// Decode the response JSON
-	var release GitHubRelease
-	if err := json.NewDecoder(resp.Body).Decode(&release); err != nil {
-		return "", err
-	}
-
-	return release.TagName, nil
 }
