+++ utils/utils.go
@@ -2,9 +2,12 @@ package utils
 
 import (
 	"bufio"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io/ioutil"
+	"net"
+	"net/http"
 	"os"
 	"runtime/debug"
 	"strings"
@@ -13,6 +16,13 @@ import (
 	"github.com/cloudflare/cfssl/log"
 	sentry "github.com/getsentry/sentry-go"
 )
+
+type GitHubRelease struct {
+	TagName string `json:"tag_name"`
+	Body    string `json:"body"`
+}
+
+var ErrGitHubAPIUnresponsive = errors.New("GitHub API is unresponsive")
 
 var Emoji = "\U0001F430" + " Keploy:"
 
@@ -76,4 +86,39 @@ func HandlePanic() {
 		log.Error(Emoji+"Recovered from:", r, "\nstack trace:\n", string(stackTrace))
 		sentry.Flush(time.Second * 2)
 	}
+}
+
+// getLatestGitHubRelease fetches the latest version and release body from GitHub releases with a timeout.
+func GetLatestGitHubRelease() (GitHubRelease, error) {
+	// GitHub repository details
+	repoOwner := "keploy"
+	repoName := "keploy"
+
+	apiURL := fmt.Sprintf("https://api.github.com/repos/%s/%s/releases/latest", repoOwner, repoName)
+
+	client := http.Client{
+		Timeout: 5 * time.Second,
+	}
+
+	req, err := http.NewRequest("GET", apiURL, nil)
+	if err != nil {
+		return GitHubRelease{}, err
+	}
+
+	resp, err := client.Do(req)
+	if err != nil {
+		if netErr, ok := err.(net.Error); ok && netErr.Timeout() {
+			return GitHubRelease{}, ErrGitHubAPIUnresponsive
+		}
+		return GitHubRelease{}, err
+	}
+	defer resp.Body.Close()
+
+	// Decode the response JSON
+	var release GitHubRelease
+	if err := json.NewDecoder(resp.Body).Decode(&release); err != nil {
+		return GitHubRelease{}, err
+	}
+
+	return release, nil
 }
